finish MLP-Mixer
